requirements.txt
streamlit
transformers
transformers[torch]
sentencepiece
diffusers
torch
colorthief
pillow
opencv-python
pandas
numpy
torch
torchvision
ultralytics
nltk
Levenshtein